Issue #164: wait function raises a TimeoutException when a process returns -1 (Windows)
We now return WAIT_TIMEOUT (0x00000102) instead of -1 when a timeout occurs. 
Collision with a process return value is still possible but less likely than 
with the original -1 value.

Also added close brackets for all if statements per coding style for the rest
of the module and to avoid potential logic bugs in the future.


git-svn-id: http://psutil.googlecode.com/svn/trunk@971 d8a29897-7e4b-0410-a74a-8bb4c295346f